requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
streamlit==1.37.1
pandas==2.2.2
requests==2.32.3
openai>=1.40.0
yfinance==0.2.29
python-dotenv==1.0.1
plotly==5.23.0
textblob==0.18.0.post0
sqlalchemy==2.0.19
reportlab==4.2.2
httpx>=0.24.1
